+++ application/views/payments/list_payments.php
@@ -55,6 +55,7 @@
     <?php $this->load->view('Templates/footer'); ?>
 </div>
 <script>
+    var permissions = <?php echo json_encode($perArr); ?>;
     $(function () {
         $('.datatable-basic').dataTable({
             autoWidth: false,
@@ -111,8 +112,12 @@
                     sortable: false,
                     render: function (data, type, full, meta) {
                         var action = '';
-                        action += '<a href="' + site_url + 'payments/edit/' + btoa(full.id) + '" class="btn border-primary text-primary-600 btn-flat btn-icon btn-rounded btn-xs" title="Edit Payment"><i class="icon-pencil3"></i></a>';
-                        action += '&nbsp;&nbsp;<a href="' + site_url + 'payments/delete/' + btoa(full.id) + '" class="btn border-danger text-danger-600 btn-flat btn-icon btn-rounded btn-xs" onclick="return confirm_alert(this)" title="Delete Payment"><i class="icon-trash"></i></a>'
+                        if ($.inArray('edit', permissions) !== -1) {
+                            action += '<a href="' + site_url + 'payments/edit/' + btoa(full.id) + '" class="btn border-primary text-primary-600 btn-flat btn-icon btn-rounded btn-xs" title="Edit Payment"><i class="icon-pencil3"></i></a>';
+                        }
+                        if ($.inArray('delete', permissions) !== -1) {
+                            action += '&nbsp;&nbsp;<a href="' + site_url + 'payments/delete/' + btoa(full.id) + '" class="btn border-danger text-danger-600 btn-flat btn-icon btn-rounded btn-xs" onclick="return confirm_alert(this)" title="Delete Payment"><i class="icon-trash"></i></a>'
+                        }
                         return action;
                     }
                 }
@@ -134,15 +139,14 @@
             confirmButtonColor: "#FF7043",
             confirmButtonText: "Yes, delete it!"
         },
-        function (isConfirm) {
-            if (isConfirm) {
-                window.location.href = $(e).attr('href');
-                return true;
-            }
-            else {
-                return false;
-            }
-        });
+                function (isConfirm) {
+                    if (isConfirm) {
+                        window.location.href = $(e).attr('href');
+                        return true;
+                    } else {
+                        return false;
+                    }
+                });
         return false;
     }
 </script>
